........
  r696501 | josh | 2008-09-17 15:35:09 -0700 (Wed, 17 Sep 2008) | 1 line

  Small tweak to NamePtg and Ref3DPtg
........


git-svn-id: https://svn.apache.org/repos/asf/poi/branches/ooxml@696831 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/record/NameRecord.java
@@ -447,7 +447,7 @@ public final class NameRecord extends Record {
 	/** gets the extern sheet number
 	 * @return extern sheet index
 	 */
-	public short getExternSheetNumber(){
+	public int getExternSheetNumber(){
 		if (field_13_name_definition.length < 1) {
 			return 0;
 		}
@@ -513,7 +513,7 @@ public final class NameRecord extends Record {
 			oldPtg = field_13_name_definition[0];
 		}
 		List temp = new ArrayList();
-		short externSheetIndex = 0;
+		int externSheetIndex = 0;
 
 		if (oldPtg.getClass() == Area3DPtg.class){
 			externSheetIndex =  ((Area3DPtg) oldPtg).getExternSheetIndex();
